Remove open methods from extensions
